requirements/base.txt
mysql-python
Django
South
unipath
distribute
python-openid
django-userena
easy-thumbnails
django-guardian
PIL
django-reversion